requirements.txt
mage-ai
pymysql
pandas
cryptography
jupyter
notebook
snowflake-connector-python
matplotlib
tabulate